# Refactor and speedup CleanFileCacheTest