--- a/Framework/PythonInterface/test/python/plugins/algorithms/CleanFileCacheTest.py
+++ b/Framework/PythonInterface/test/python/plugins/algorithms/CleanFileCacheTest.py
@@ -6,10 +6,10 @@
 # SPDX - License - Identifier: GPL - 3.0 +
 # pylint: disable=invalid-name,too-many-public-methods,too-many-arguments,multiple-statements
 import unittest
+
 from mantid.kernel import *
 from mantid.api import *
 from mantid.simpleapi import CreateCacheFilename
-from typing import List
 
 import datetime, os, tempfile, glob, shutil
 
@@ -33,61 +33,52 @@ class CleanFileCache(unittest.TestCase):
             ]
         ]
         for filepath in cls._non_cache_filepaths:
-            touch(filepath)
+            _write_file(filepath)
+
+        cls._cache_filepath1, _ = CreateCacheFilename(CacheDir=cls._cache_root, OtherProperties=["A=first"])
+        cls._cache_filepath2, _ = CreateCacheFilename(CacheDir=cls._cache_root, OtherProperties=["B=second"])
+        cls._cache_filepath3, _ = CreateCacheFilename(CacheDir=cls._cache_root, OtherProperties=["C=third"])
 
     @classmethod
     def tearDownClass(cls) -> None:
         shutil.rmtree(cls._cache_root)
 
-    def test2(self):
-        """CleanFileCache: 'normal' files with 39 and 41-character filenames etc"""
-        self._create_file_in_cache(["A=1"])
-        self._create_file_in_cache(["B='silly'"])
+    def test_clean_cache_will_not_clear_non_cache_files(self):
+        _write_file(self._cache_filepath1)
+        _write_file(self._cache_filepath2)
 
-        execute_clean_cache(self._cache_root, 0)
+        _execute_clean_cache(self._cache_root, 0)
 
         files_remained = glob.glob(os.path.join(self._cache_root, "*"))
-
         self.assertEqual(set(files_remained), set(self._non_cache_filepaths))
 
-    def test3(self):
-        """CleanFileCache: "age" parameter"""
+    def test_clean_cache_will_not_clear_a_file_older_than_an_age_limit(self):
         age = 10
 
-        cache1 = self._create_file_in_cache(["A=newer"], age - 1)
-        self._create_file_in_cache(["B=rightonedge"], age)
-        self._create_file_in_cache(["C=old"], age + 1)
+        _write_file(self._cache_filepath1, age - 1)
+        _write_file(self._cache_filepath2, age)
+        _write_file(self._cache_filepath3, age + 1)
 
-        execute_clean_cache(self._cache_root, age)
+        _execute_clean_cache(self._cache_root, age)
 
         files_remained = glob.glob(os.path.join(self._cache_root, "*"))
-        self.assertEqual(set(files_remained), set(self._non_cache_filepaths + [cache1]))
-
-    def _create_file_in_cache(self, properties: List[str], days_before=None):
-        filepath, _ = CreateCacheFilename(CacheDir=self._cache_root, OtherProperties=properties)
-        touch(filepath)
-
-        if days_before is not None:
-            t = computeTime(days_before)
-            os.utime(filepath, (t, t))
-        return filepath
+        self.assertEqual(set(files_remained), set(self._non_cache_filepaths + [self._cache_filepath1]))
 
 
-def execute_clean_cache(cache_root: str, age: int):
+def _execute_clean_cache(cache_root: str, age: int):
     from mantid.simpleapi import CleanFileCache
 
     CleanFileCache(CacheDir=cache_root, AgeInDays=age)
 
 
-def computeTime(days_before: int):
-    "compute time as float of the time at n=daysbefore days before today"
-    return TIME_3PM - days_before * 24 * 60 * 60
-
-
-def touch(f):
-    with open(f, "w") as stream:
+def _write_file(filepath: str, days_before: int = None) -> None:
+    with open(filepath, "w") as stream:
         stream.write("\n")
-    return
+
+    if days_before is not None:
+        # Convert to seconds before 3pm today
+        t = TIME_3PM - days_before * 24 * 60 * 60
+        os.utime(filepath, (t, t))
 
 
 if __name__ == "__main__":